+++ core-ng/src/main/java/core/framework/internal/web/sse/ServerSentEventCloseHandler.java
@@ -5,24 +5,18 @@ import core.framework.internal.log.ActionLog;
 import core.framework.internal.log.LogManager;
 import io.undertow.server.ExchangeCompletionListener;
 import io.undertow.server.HttpServerExchange;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.List;
-import java.util.Map;
 
 class ServerSentEventCloseHandler<T> implements ExchangeCompletionListener {
-    private static final Logger LOGGER = LoggerFactory.getLogger(ServerSentEventCloseHandler.class);
-    final ServerSentEventContextImpl<T> context;
+    final ChannelSupport<T> support;
     private final LogManager logManager;
     private final ChannelImpl<T> channel;
-    private final Map<String, String> logContext;
 
-    ServerSentEventCloseHandler(LogManager logManager, ChannelImpl<T> channel, ServerSentEventContextImpl<T> context, Map<String, String> logContext) {
+    ServerSentEventCloseHandler(LogManager logManager, ChannelImpl<T> channel, ChannelSupport<T> support) {
         this.logManager = logManager;
         this.channel = channel;
-        this.context = context;
-        this.logContext = logContext;
+        this.support = support;
     }
 
     @Override
@@ -33,17 +27,17 @@ class ServerSentEventCloseHandler<T> implements ExchangeCompletionListener {
             try {
                 actionLog.action("sse:" + exchange.getRequestPath() + ":close");
                 actionLog.context("channel", channel.id);
-                LOGGER.debug("refId={}", channel.refId);
                 List<String> refIds = List.of(channel.refId);
                 actionLog.refIds = refIds;
                 actionLog.correlationIds = refIds;
 
-                for (Map.Entry<String, String> entry : logContext.entrySet()) {
-                    actionLog.context(entry.getKey(), entry.getValue());
-                }
+                actionLog.context("client_id", channel.clientIP);
+                if (channel.traceId != null) actionLog.context("trace_id", channel.traceId);
+
+                support.listener.onClose(channel);
 
                 if (!channel.groups.isEmpty()) actionLog.context("group", channel.groups.toArray());
-                context.remove(channel);
+                support.context.remove(channel);
                 channel.shutdown();
             } catch (Throwable e) {
                 logManager.logError(e);