+++ core-ng/src/main/java/core/framework/web/sse/ChannelListener.java
@@ -4,4 +4,7 @@ import core.framework.web.Request;
 
 public interface ChannelListener<T> {
     void onConnect(Request request, Channel<T> channel, String lastEventId);
+
+    default void onClose(Channel<T> channel) {
+    }
 }